requirements.txt
requests
pyyaml
pandas
pyarrow
pytest
